cloudBackend/CloudFunction/MachineLearningModule/sentimentAnalysis/requirements.txt
# Function dependencies, for example:
# package>=version
argparse
google-cloud-language
six